+++ src/BootstrapVue.ts
@@ -84,6 +84,8 @@ export {default as BTable} from './components/BTable/BTable.vue'
 export {default as BTabs} from './components/BTabs/BTabs.vue'
 export {default as BTransition} from './components/BTransition/BTransition.vue'
 export {default as BToast} from './components/BToast/BToast.vue'
+export {default as BToaster} from './components/BToast/BToaster.vue'
+export {ToastPlugin} from './components/BToast/'
 
 const plugin: Plugin = {
   // TODO: use options in the future
