+++ src/components/BToast/index.ts
@@ -1,0 +1,5 @@
+import BToast from './BToast.vue'
+import BToaster from './BToaster.vue'
+import ToastPlugin from './plugin'
+
+export {ToastPlugin, BToast, BToaster}
