+++ src/components/index.ts
@@ -72,7 +72,8 @@ import BSpinner from './BSpinner.vue'
 import BTab from './BTabs/BTab.vue'
 import BTable from './BTable/BTable.vue'
 import BTabs from './BTabs/BTabs.vue'
-import BToast from './BToast/BToast.vue'
+import {BToast, BToaster, ToastPlugin} from './BToast'
+
 import BToastContainer from './BToast/BToaster.vue'
 import BTransition from './BTransition/BTransition.vue'
 
@@ -151,7 +152,9 @@ export default {
   BTab,
   BTable,
   BToast,
+  BToaster,
   BToastContainer,
   BTabs,
   BTransition,
+  ToastPlugin,
 }
