--- src/plugins/index.ts
@@ -1,5 +1,0 @@
-import BToastPlugin from './BToast'
-
-export default {
-  BToastPlugin,
-}
